Renaming of several items
ShowFeedback.showTest -> ShowFeedback.showTestResults
SptCheckForOutput -> SptRunTestSuite
SptRunTestSuite -> SptShowTestSuiteResults
TestCaseRun -> TestCaseResult
TestSuiteRun -> TestSuiteResult
MultiTestSuiteRun -> TestResults
