+++ core/spoofax.core/src/main/java/mb/spoofax/core/language/command/ShowFeedback.java
@@ -3,7 +3,7 @@ package mb.spoofax.core.language.command;
 import mb.common.region.Region;
 import mb.common.util.ADT;
 import mb.resource.ResourceKey;
-import mb.spoofax.core.language.testrunner.MultiTestSuiteRun;
+import mb.spoofax.core.language.testrunner.TestResults;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.Serializable;
@@ -16,7 +16,7 @@ public abstract class ShowFeedback implements Serializable {
 
         R showText(String text, String name, @Nullable Region region);
 
-        R showTests(MultiTestSuiteRun testResults, @Nullable Region region);
+        R showTestResults(TestResults testResults, @Nullable Region region);
     }
 
     @SuppressWarnings("ConstantConditions")
@@ -39,7 +39,7 @@ public abstract class ShowFeedback implements Serializable {
         return ShowFeedbacks.showText(text, name, null);
     }
 
-    public static ShowFeedback showTests(MultiTestSuiteRun testResults) { return ShowFeedbacks.showTests(testResults, null); }
+    public static ShowFeedback showTestResults(TestResults testResults) { return ShowFeedbacks.showTestResults(testResults, null); }
 
     public abstract <R> R match(Cases<R> cases);
 
@@ -64,7 +64,7 @@ public abstract class ShowFeedback implements Serializable {
         return Optional.ofNullable(region);
     }
 
-    public Optional<MultiTestSuiteRun> getTestResults() {
+    public Optional<TestResults> getTestResults() {
         return ShowFeedbacks.getTestResults(this);
     }
