+++ core/spoofax.eclipse/src/main/java/mb/spoofax/eclipse/testrunner/TestRunLabelProvider.java
@@ -9,8 +9,8 @@ import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.widgets.Display;
-import mb.spoofax.core.language.testrunner.TestCaseRun;
-import mb.spoofax.core.language.testrunner.TestSuiteRun;
+import mb.spoofax.core.language.testrunner.TestCaseResult;
+import mb.spoofax.core.language.testrunner.TestSuiteResult;
 
 /**
  * Returns the labels for the suites and test cases.
@@ -19,12 +19,12 @@ public class TestRunLabelProvider extends LabelProvider
     implements ITableLabelProvider, ITableFontProvider, ITableColorProvider {
 
     @Override public String getText(Object element) {
-        if(element instanceof TestSuiteRun) {
-            TestSuiteRun tsr = (TestSuiteRun) element;
+        if(element instanceof TestSuiteResult) {
+            TestSuiteResult tsr = (TestSuiteResult) element;
             int failed = tsr.numFailed();
             return failed == 0 ? tsr.name : String.format("%s (%d failed)", tsr.name, failed);
-        } else if(element instanceof TestCaseRun) {
-            TestCaseRun tcr = (TestCaseRun) element;
+        } else if(element instanceof TestCaseResult) {
+            TestCaseResult tcr = (TestCaseResult) element;
             String lbl = tcr.description;
             if(tcr.result() != null) {
                 lbl = lbl + " (" + String.format("%.2f", tcr.duration() / 1000.0) + "s)";
@@ -38,8 +38,8 @@ public class TestRunLabelProvider extends LabelProvider
     }
 
     @Override @Nullable public Color getForeground(Object element, int columnIndex) {
-        if(element instanceof TestSuiteRun) {
-            TestSuiteRun tsr = (TestSuiteRun) element;
+        if(element instanceof TestSuiteResult) {
+            TestSuiteResult tsr = (TestSuiteResult) element;
             if(tsr.messages != null && !tsr.messages.containsError()) {
                 // use default color;
                 return null;
@@ -47,8 +47,8 @@ public class TestRunLabelProvider extends LabelProvider
                 // use red
                 return new Color(Display.getCurrent(), 159, 63, 63);
             }
-        } else if(element instanceof TestCaseRun) {
-            TestCaseRun tcr = (TestCaseRun) element;
+        } else if(element instanceof TestCaseResult) {
+            TestCaseResult tcr = (TestCaseResult) element;
             if(tcr.result() != null) {
                 return !tcr.result().containsError() ? new Color(Display.getCurrent(), 10, 100, 10)
                     : new Color(Display.getCurrent(), 159, 63, 63);
